requirements.txt
numpy
six
tqdm>4.11.1
msgpack-python>0.4.0
msgpack-numpy>=0.3.9
pyzmq>=16